// src/app/page/master/pegawai-sk-gaji/pegawai-sk-gaji.interface.ts
import { EventEmitter } from '@angular/core';

export interface PegawaiSkGaji {
    kode?: any;
    pegawai?: number;
    kdKomponenHarga?: string;
    kdOperatorRate?: string;
    keterangan?: string;
    hargaSatuan?: string;
    persenHargaSatuan?: string;
    qtyProduk?: string;
    status1?: string;
    kdMetodePembayaran?: string;
    kdMataUang?: string;
    statusEnabled?: boolean;
    noRec?: string;
    label?: any;
    value?: any;
    namaStatus?: string;
    id?: any;
}